requirements.txt

fastapi==0.109.2
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pyravendb==5.0.0.5
python-multipart==0.0.9
ravendb==5.2.5.post1
requests==2.31.0
urllib3==2.2.0
uvicorn==0.27.1
